+++ core/src/components/datetime/utils/data.ts
@@ -411,7 +411,7 @@ export const getYearColumnData = (
     const maxYear = maxParts?.year ?? year;
     const minYear = minParts?.year ?? year - 100;
 
-    for (let i = maxYear; i >= minYear; i--) {
+    for (let i = minYear; i <= maxYear; i++) {
       processedYears.push(i);
     }
   }